fix: Phase 1-4 最終レビュー指摘修正
AudioClipBuilder 安全性修正:
- BuildAudioClip の SetData 失敗時に作成済み AudioClip を
  DestroyImmediate で確実にクリーンアップ（float[]/NativeArray両版）

テスト共通ヘルパー集約:
- TestHelpers.cs 新規作成（CreateMinimalPhonemeIdMap, CreateValidatedConfig,
  SynchronousProgress<T> を3ファイルから集約）
- SplitInferenceOrchestratorTests, TTSSynthesisOrchestratorTests,
  PiperTTSSynthesizeWithTimingTests の重複メソッド除去（-173行, +74行）

P3-1 チケット文書修正:
- セクション2.3/2.5/2.7 に実装ノート追加
  （durations結合ロジックがPhase 3レビューで除去された経緯を明記）
