+++ Assets/uPiper/Tests/Editor/AudioGeneration/SplitInferenceOrchestratorTests.cs
@@ -17,26 +17,6 @@ namespace uPiper.Tests.Editor.AudioGeneration
         private StubInferenceAudioGenerator _stubGenerator;
         private SplitInferenceOrchestrator _orchestrator;
 
-        /// <summary>
-        /// テスト用の最小限 phonemeIdMap。
-        /// PhonemeSilenceProcessor.SplitAtPhonemeSilence が
-        /// phonemeSilence のキーを phonemeIdMap で逆引きするために使用。
-        /// </summary>
-        private static Dictionary<string, int[]> CreateMinimalPhonemeIdMap()
-        {
-            return new Dictionary<string, int[]>
-            {
-                ["_"] = new[] { 0 },  // PAD（沈黙トークンとしても使用）
-                ["^"] = new[] { 1 },  // BOS
-                ["$"] = new[] { 2 },  // EOS
-                ["a"] = new[] { 3 },
-                ["i"] = new[] { 4 },
-                ["u"] = new[] { 5 },
-                ["k"] = new[] { 8 },
-                ["o"] = new[] { 7 },
-            };
-        }
-
         [SetUp]
         public void SetUp()
         {
@@ -57,7 +37,7 @@ namespace uPiper.Tests.Editor.AudioGeneration
             // phonemeSilence に "#" を設定するが、phonemeIds には "#" の ID が無いため分割されない
             var phonemeIds = new[] { 1, 3, 4, 5, 2 }; // ^, a, i, u, $
             var phonemeSilence = new Dictionary<string, float> { ["#"] = 0.5f };
-            var phonemeIdMap = CreateMinimalPhonemeIdMap();
+            var phonemeIdMap = TestHelpers.CreateMinimalPhonemeIdMap();
             phonemeIdMap["#"] = new[] { 99 }; // phonemeSilence のキー用（phonemeIds に含まれない）
 
             // Act
@@ -93,7 +73,7 @@ namespace uPiper.Tests.Editor.AudioGeneration
             // "_" の位置で分割される → 句1: [1,3,0], 句2: [8,7,2]
             var phonemeIds = new[] { 1, 3, 0, 8, 7, 2 };
             var phonemeSilence = new Dictionary<string, float> { ["_"] = 0.5f };
-            var phonemeIdMap = CreateMinimalPhonemeIdMap();
+            var phonemeIdMap = TestHelpers.CreateMinimalPhonemeIdMap();
 
             // 0.5秒 × 22050Hz = 11025 サンプルの無音が挿入される
             var expectedSilenceSamples = (int)(0.5f * 22050);
@@ -141,7 +121,7 @@ namespace uPiper.Tests.Editor.AudioGeneration
             // GenerateWithSilenceSplitAsync は PhonemeIds が空の Phrase をスキップする
             var phonemeIds = Array.Empty<int>();
             var phonemeSilence = new Dictionary<string, float> { ["_"] = 0.5f };
-            var phonemeIdMap = CreateMinimalPhonemeIdMap();
+            var phonemeIdMap = TestHelpers.CreateMinimalPhonemeIdMap();
 
             // Act
             var output = await _orchestrator.GenerateWithSilenceSplitAsync(
@@ -185,7 +165,7 @@ namespace uPiper.Tests.Editor.AudioGeneration
                 0, 0, 0   // phoneme 5: a1=0, a2=0, a3=0
             };
             var phonemeSilence = new Dictionary<string, float> { ["_"] = 0.5f };
-            var phonemeIdMap = CreateMinimalPhonemeIdMap();
+            var phonemeIdMap = TestHelpers.CreateMinimalPhonemeIdMap();
 
             // Act
             var output = await _orchestrator.GenerateWithSilenceSplitAsync(
@@ -220,7 +200,7 @@ namespace uPiper.Tests.Editor.AudioGeneration
             // → 句1: [1,3,0], 句2: [8,0], 句3: [7,2]
             var phonemeIds = new[] { 1, 3, 0, 8, 0, 7, 2 };
             var phonemeSilence = new Dictionary<string, float> { ["_"] = 0.2f };
-            var phonemeIdMap = CreateMinimalPhonemeIdMap();
+            var phonemeIdMap = TestHelpers.CreateMinimalPhonemeIdMap();
 
             var reportedValues = new List<float>();
             var progress = new Progress<float>(v => reportedValues.Add(v));
@@ -259,7 +239,7 @@ namespace uPiper.Tests.Editor.AudioGeneration
             // Arrange: 沈黙トークンを含まない phonemeIds → 分割なし → 1句
             var phonemeIds = new[] { 1, 3, 4, 5, 2 }; // ^, a, i, u, $
             var phonemeSilence = new Dictionary<string, float> { ["#"] = 0.5f };
-            var phonemeIdMap = CreateMinimalPhonemeIdMap();
+            var phonemeIdMap = TestHelpers.CreateMinimalPhonemeIdMap();
             phonemeIdMap["#"] = new[] { 99 };
 
             // SynchronousProgress は Report を同期的に呼び出すため、
@@ -296,7 +276,7 @@ namespace uPiper.Tests.Editor.AudioGeneration
             // Arrange
             var phonemeIds = Array.Empty<int>();
             var phonemeSilence = new Dictionary<string, float> { ["_"] = 0.5f };
-            var phonemeIdMap = CreateMinimalPhonemeIdMap();
+            var phonemeIdMap = TestHelpers.CreateMinimalPhonemeIdMap();
 
             var reportedValues = new List<float>();
             var progress = new SynchronousProgress<float>(v => reportedValues.Add(v));
@@ -327,7 +307,7 @@ namespace uPiper.Tests.Editor.AudioGeneration
             // Arrange: 3句に分割
             var phonemeIds = new[] { 1, 3, 0, 8, 0, 7, 2 };
             var phonemeSilence = new Dictionary<string, float> { ["_"] = 0.2f };
-            var phonemeIdMap = CreateMinimalPhonemeIdMap();
+            var phonemeIdMap = TestHelpers.CreateMinimalPhonemeIdMap();
 
             var reportedValues = new List<float>();
             var progress = new SynchronousProgress<float>(v => reportedValues.Add(v));
@@ -371,7 +351,7 @@ namespace uPiper.Tests.Editor.AudioGeneration
             // Arrange: 短い phonemeIds（5要素 < MinPhonemeIds=40）
             var phonemeIds = new[] { 1, 5, 6, 7, 2 }; // ^, u, ?, ?, $
             var phonemeSilence = new Dictionary<string, float>(); // 空 → 分割なし
-            var phonemeIdMap = CreateMinimalPhonemeIdMap();
+            var phonemeIdMap = TestHelpers.CreateMinimalPhonemeIdMap();
 
             // StubをShortTextMitigatingGeneratorでラップ
             var wrappedGenerator = new ShortTextMitigatingGenerator(_stubGenerator);
@@ -410,7 +390,7 @@ namespace uPiper.Tests.Editor.AudioGeneration
             _stubGenerator.DurationsToReturn = new float[] { 1f, 2f, 3f };
             var phonemeIds = new[] { 1, 3, 0, 8, 7, 2 }; // ^,a,_,k,o,$
             var phonemeSilence = new Dictionary<string, float> { ["_"] = 0.5f };
-            var phonemeIdMap = CreateMinimalPhonemeIdMap();
+            var phonemeIdMap = TestHelpers.CreateMinimalPhonemeIdMap();
 
             // Act
             var output = await _orchestrator.GenerateWithSilenceSplitAsync(
@@ -439,7 +419,7 @@ namespace uPiper.Tests.Editor.AudioGeneration
             _stubGenerator.AudioDataToReturn = CreateFixedAudioData(30, 0.3f);
             var phonemeIds = new[] { 1, 3, 0, 8, 0, 7, 2 }; // ^,a,_,k,_,o,$
             var phonemeSilence = new Dictionary<string, float> { ["_"] = 0.2f };
-            var phonemeIdMap = CreateMinimalPhonemeIdMap();
+            var phonemeIdMap = TestHelpers.CreateMinimalPhonemeIdMap();
 
             var expectedSilenceSamples = (int)(0.2f * 22050);
 
@@ -475,7 +455,7 @@ namespace uPiper.Tests.Editor.AudioGeneration
             _stubGenerator.DurationsToReturn = new float[] { 1f, 2f, 3f, 4f, 5f };
             var phonemeIds = new[] { 1, 3, 4, 5, 2 }; // ^,a,i,u,$
             var phonemeSilence = new Dictionary<string, float> { ["#"] = 0.5f };
-            var phonemeIdMap = CreateMinimalPhonemeIdMap();
+            var phonemeIdMap = TestHelpers.CreateMinimalPhonemeIdMap();
             phonemeIdMap["#"] = new[] { 99 };
 
             // Act
@@ -504,7 +484,7 @@ namespace uPiper.Tests.Editor.AudioGeneration
             _stubGenerator.DurationsToReturn = null;
             var phonemeIds = new[] { 1, 3, 0, 8, 7, 2 };
             var phonemeSilence = new Dictionary<string, float> { ["_"] = 0.5f };
-            var phonemeIdMap = CreateMinimalPhonemeIdMap();
+            var phonemeIdMap = TestHelpers.CreateMinimalPhonemeIdMap();
 
             // Act
             var output = await _orchestrator.GenerateWithSilenceSplitAsync(
@@ -536,7 +516,7 @@ namespace uPiper.Tests.Editor.AudioGeneration
             // Arrange: 空 phonemeIds
             var phonemeIds = Array.Empty<int>();
             var phonemeSilence = new Dictionary<string, float> { ["_"] = 0.5f };
-            var phonemeIdMap = CreateMinimalPhonemeIdMap();
+            var phonemeIdMap = TestHelpers.CreateMinimalPhonemeIdMap();
             _stubGenerator.DurationsToReturn = new float[] { 1f, 2f };
 
             // Act
@@ -568,7 +548,7 @@ namespace uPiper.Tests.Editor.AudioGeneration
             _stubGenerator.AudioDataToReturn = CreateFixedAudioData(50, 0.5f);
             var phonemeIds = new[] { 1, 3, 0, 8, 7, 2 };
             var phonemeSilence = new Dictionary<string, float> { ["_"] = 0.5f };
-            var phonemeIdMap = CreateMinimalPhonemeIdMap();
+            var phonemeIdMap = TestHelpers.CreateMinimalPhonemeIdMap();
 
             var expectedSilenceSamples = (int)(0.5f * 22050);
 
@@ -606,25 +586,5 @@ namespace uPiper.Tests.Editor.AudioGeneration
             return data;
         }
 
-        /// <summary>
-        /// テスト用の同期的 IProgress 実装。
-        /// <see cref="Progress{T}"/> は SynchronizationContext.Post を使用するため、
-        /// EditMode テストではコールバックの発火タイミングが不定。
-        /// このクラスは Report() を同期的に呼び出す。
-        /// </summary>
-        private sealed class SynchronousProgress<T> : IProgress<T>
-        {
-            private readonly Action<T> _handler;
-
-            public SynchronousProgress(Action<T> handler)
-            {
-                _handler = handler ?? throw new ArgumentNullException(nameof(handler));
-            }
-
-            public void Report(T value)
-            {
-                _handler(value);
-            }
-        }
     }
 }
