+++ docs/tickets/P3-1.md
@@ -257,6 +257,13 @@ public async Task<InferenceOutput> GenerateWithSilenceSplitAsync(
 }
 ```
 
+> **実装ノート（2026-04-12 Phase 3 レビュー修正）**:
+> 上記のdurations結合ロジックはPhase 3レビューで除去された。
+> TTSSynthesisOrchestrator.CalculateTimings が句分割使用時にタイミング計算をスキップする（`null` を返す）ため、
+> 結合されたdurationsは使用されずに破棄されることが判明した。
+> 最終実装では `return new InferenceOutput(combinedAudio, default);` として常にdurationsなしで返却する。
+> 将来の句単位タイミング対応では `SplitInferenceResult` 型を導入し、句ごとのメタデータを保持する設計を予定。
+
 ### 2.4 P2-1 暫定コードの除去
 
 P2-1 では SplitInferenceOrchestrator 内で `InferenceOutput` から Audio のみ抽出し Durations を破棄する暫定実装となっていた。P3-1 ではこの暫定コードを完全に置き換え、durations 結合ロジックを実装する。
@@ -294,6 +301,10 @@ finally
 
 TTSSynthesisOrchestrator（P3-2）では、句分割を使用した場合に `TimingCalculator` の結果に句間無音のオフセットを加算するロジックが必要になる。ただし、P3-2 の設計判断として「句分割時はタイミング情報を提供しない（HasTimings=false）」とする選択肢もあり得る。この判断は P3-2 に委ねる。
 
+> **補足**: Phase 3レビューの結果、上記の議論に基づきSplitInferenceOrchestratorではdurationsを結合せず、
+> 常に `default` durationsを返す実装に確定した。句分割時のタイミング計算はTTSSynthesisOrchestratorで
+> スキップされる（CalculateTimings が `usedSilenceSplit == true` で `null` を返す）。
+
 ### 2.6 Durations 結合の正当性
 
 各句は独立した推論として実行され、それぞれの durations は句内の音素に対するフレーム持続時間を表す。句ごとの durations を単純結合した配列は、元の（分割前の）phonemeIds に対応する durations 配列と等価になる。
@@ -314,6 +325,9 @@ TTSSynthesisOrchestrator（P3-2）では、句分割を使用した場合に `Ti
 `hasDurations` フラグは「いずれかの句が durations を持っていた」ことを示す。全句が durations を持たない場合（`hasDurations == false`）、結合 durations は `default(NativeArray<float>)` となる。
 
 一部の句のみ durations を持つ不整合な状態は想定しない。`SupportsDurations` はモデル単位のプロパティであり、同一推論セッション内で変化しないため、全句が同一の durations サポート状態を持つ。
+
+> **補足**: Phase 3レビュー修正により、`hasDurations` フラグとdurations結合ロジックは除去された。
+> SplitInferenceOrchestratorは常に `HasDurations == false` の InferenceOutput を返す。
 
 ### 2.8 ログメッセージ
